Increase size of response message buffer
We were experiencing a lack of function due to an overly long 'output'
element coupled with quite a few overlay mounts, which lead to the
buffer not even containing the network section.

A simple doubling in buffer size fixes this problem, but since the
container information can be arbitrarily long, this is just a
jury-rigged fix.